app/requirements.txt
#postgres database adapater
psycopg2==2.6.1

#django framework itself
Django==1.9

#rest framework
djangorestframework==3.3.1

#yaml for fixtures
PyYAML==3.11

#appserver
gunicorn==19.4.1

elasticsearch>=2.0.0,<3.0.0
textblob==0.11.1
tweepy==3.5.0

celery[redis]
feedparser==5.2.1